fix(bser)+feat(core): fix bser serialization and implemented device hash

+++ actor-ios/src/im/actor/model/droidkit/bser/DataInput.h
@@ -42,7 +42,7 @@
 
 - (IOSByteArray *)readBytesWithInt:(jint)count;
 
-- (jlong)readVarInt32;
+- (jint)readVarInt32;
 
 - (jlong)readVarInt;
